Edited according to the recommendations

+++ src/com/epam/entities/CasualCar.java
@@ -1,7 +1,9 @@
 package com.epam.entities;
 
-public abstract class CasualCar extends PassengerCar{
-    public CasualCar(int purchaseCost, int yearFromPurchase) {
+import java.math.BigDecimal;
+
+public abstract class CasualCar extends PassengerCar {
+    public CasualCar(BigDecimal purchaseCost, int yearFromPurchase) {
         super(purchaseCost, yearFromPurchase);
     }
 }
